+++ docs/CHANGELOG.md
@@ -2,6 +2,13 @@
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
+## [1.0.2] - 2024-10-10
+### Added
+- Generate & Validate Cache (cache/generate_validate_cache.bat) - Leveraging the scripts in the cache directory to automate UJITSO caching and validation.
+- Run Variants (cache/run_variants.py) - Run all variants in a stage awaiting the stage to be ready between each variant being set. Works both with caching graph and generic Python as back up triggering variants found.
+- Copy Configurator (cache/copy_configurator.py) - Copy the configurator to another folder on the local disk.
+- Validate Log (cache/validate_log.py) - Find UJITSO errors in log file.
+  
 ## [1.0.1] - 2024-08-15
 ### Added
 - csv_to_json.py dataset for Concept Car asset